REF: Re-organized Foyer module
