+++ openff/interchange/foyer/__init__.py
@@ -1,0 +1,1 @@
+"""Foyer compatibility."""
