+++ setup.cfg
@@ -15,8 +15,11 @@ max-line-length = 119
 ignore = E203,B028
 per-file-ignores =
     openff/interchange/tests/unit_tests/test_types.py:F821
-    openff/interchange/smirnoff/_*:F821
+    openff/interchange/smirnoff/_nonbonded.py:F821
+    openff/interchange/smirnoff/_gbsa.py:F821
+    openff/interchange/foyer/_nonbonded.py:F821
     openff/interchange/components/*:F821
+    openff/interchange/components/foyer.py:F401
     openff/interchange/__init__.py:F401
     openff/interchange/interop/gromacs/__init__.py:F401
     openff/interchange/tests/data/*:INP001
