+++ stubs/foyer/forcefield.pyi
@@ -10,5 +10,7 @@ class Forcefield(object):
     ) -> Dict[str, float]: ...
     @staticmethod
     def get_generator(ff: Forcefield, gen_type: Any) -> Any: ...
-    def lj14scale(self) -> int: ...
-    def coulomb14scale(self) -> int: ...
+    @property
+    def lj14scale(self) -> float: ...
+    @property
+    def coulomb14scale(self) -> float: ...
